.vimrc: remove some unused plugins
I think pylint and/or pep8 wouldn't shut up when building C code. Really
annoying. I have not used them for a long time anyway.

I don't remember what it is with Superior_Haskell_Interaction_Mode_SHIM,
but I've had it disabled for a long time. So just commit it.