.gitignore: remove *.pyc pattern
Although it's convenient, I think it's wrong to have this as a global
pattern. If *.pyc files are generated inside a repo, it's due to some
program from the repo being run. Thus, the ignore rule should be in that
repo.

What should be in the global ignore file though is files that are
generated outside of repo workflow, such as (arbitrary) editor save
files etc.